+++ CHANGELOG.md
@@ -21,6 +21,8 @@
 
 ### Enhancements
 
+* [ObjectServer] Adding user lookup API for admin (#4828).
+
 ### Bug Fixes
 
 ### Internal